+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Issue #421: Training on trajectory sampling
 - Issue #423: Manage continuing task only
 
 ## [0.16.1] 2024-05-05